requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
docx2txt
docxtpl==0.16.7
pytesseract==0.3.10
opencv-python==4.8.1.78
pdfplumber==0.10.3
pymupdf==1.24.9
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
Pillow==10.1.0
spacy==3.7.2
